Features

+++ app/Console/Jobs/SendResetPassword.php
@@ -28,7 +28,7 @@ class SendResetPassword implements QueueInterface
      */
     public function handle()
     {
-        if (env('APP_ENV') == 'local' || env('SEND_EMAIL_ON_LOCAL') != 'true') {
+        if (env('APP_ENV') == 'local' && env('SEND_EMAIL_ON_LOCAL') != 'true') {
             return $this->delete();
         }
         try {
@@ -40,7 +40,7 @@ class SendResetPassword implements QueueInterface
 
             $this->delete();
         } catch (Exception $e) {
-            logger(storage_path()."logs/email.log")->error('Caught a ' . get_class($e) . ': ' . $e->getMessage(), $e->getTrace());
+            logger(storage_path("logs/email.log"))->error('Caught a ' . get_class($e) . ': ' . $e->getMessage(), $e->getTrace());
         }
     }
 }
